Changing up --output creation logic
When --list-file is included, we want to do the output file differently.

For list file mode, we want a folder or stdout.
For single file mode, we want a file or stdout.

If there is ever an existing results file, we should error out before transcribing.

Therfore outputWriters are created upfront and added to the list of files to transcribe.
This will be reworked in the next commit.